backend/requirements/base.txt
argon2-cffi==19.1.0
autopep8==1.4.4
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.12.3
Django==3.0.0
django-cors-headers==3.1.1
django-discover-runner==1.0
django-enum-choices==2.1.1
django-filter==2.2.0
django-graphql-jwt==0.2.2
django-redis==4.10.0
django-role-permissions==2.2.1
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
Pillow==6.1.0
promise==2.2.1
psycopg2==2.8.3
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pytz==2019.2
redis==3.3.8
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
django-crontab==0.7.1
celery==4.3
django-cron==0.5.1
djangorestframework==3.10.3
